requirements.txt
bottle==0.12.23
cefpython3==66.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
clr-loader==0.2.5
colorama==0.4.6
Flask==2.2.2
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
numpy==1.24.1
opencv-python==4.7.0.68
pip==21.3.1
proxy-tools==0.1.0
pycparser==2.21
pygame==2.1.2
pythonnet==3.0.1
pywebview==4.0.1
requests==2.28.2
setuptools==60.2.0
urllib3==1.26.14
Werkzeug==2.2.2
wheel==0.37.1
zipp==3.12.0
